+++ enlighten/ui/HelpFeature.py
@@ -56,6 +56,7 @@ class HelpFeature:
             Ctrl-R take/clear Reference
             Ctrl-S Save measurement
             Ctrl-T enter integration Time
+            Ctrl-X toggle eXpert mode
             
             Ctrl-Left  move cursor
             Ctrl-Right move cursor"""))